No commit message

+++ scripts/r/linux/install_ssh_server.sh
@@ -1,12 +1,25 @@
-sudo apt-get install openssh-server -y
-sudo systemctl enable ssh
+set -e
+set -x
+if [[ -f "/etc/debian_version" ]]; then
+    sudo apt-get install openssh-server -y
+    sudo systemctl enable ssh
+elif [[ -f "/etc/arch-release" ]]; then
+    sudo pacman -S --noconfirm --needed openssh
+    sudo systemctl enable sshd
+fi
 
-sudo ufw allow ssh
+if [[ -x "$(command -v ufw)" ]]; then
+    sudo ufw allow ssh
+fi
 
-sudo sed -i 's/PasswordAuthentication no/PasswordAuthentication yes/g' /etc/ssh/sshd_config
+sudo sed -E -i 's/#?PasswordAuthentication (yes|no)/PasswordAuthentication yes/g' /etc/ssh/sshd_config
 sudo sed -zi '/AllowUsers ross/!s/$/\nAllowUsers ross/' /etc/ssh/sshd_config
-sudo sed -i 's/#\?PermitRootLogin [^\n]\+/PermitRootLogin no/g' /etc/ssh/sshd_config
+sudo sed -E -i 's/#?PermitRootLogin (yes|no)/PermitRootLogin no/g' /etc/ssh/sshd_config
 
 sudo ssh-keygen -A # generate host key if does not exist
 
-sudo systemctl restart ssh.service
+if [[ -f "/etc/debian_version" ]]; then
+    sudo systemctl restart ssh.service
+elif [[ -f "/etc/arch-release" ]]; then
+    sudo systemctl restart sshd.service
+fi
